set a separate searchsuggestions component
