+++ src/components/ActionBar/SearchInput.js
@@ -1,13 +1,18 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { lazy, useCallback, useEffect, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
-import { ModalView, MongoDbLegalDisclosure, PoweredByAtlasVectorSearch, useChatbotContext } from 'mongodb-chatbot-ui';
+import styled from '@emotion/styled';
 import { css, cx } from '@leafygreen-ui/emotion';
 import { palette } from '@leafygreen-ui/palette';
-import { SearchInput as LGSearchInput, SearchResult } from '@leafygreen-ui/search-input';
-import { theme } from '../../theme/docsTheme';
+import { SearchInput as LGSearchInput } from '@leafygreen-ui/search-input';
+import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
 import debounce from '../../utils/debounce';
 import { isBrowser } from '../../utils/is-browser';
-import { ShortcutIcon, SparkleIcon } from './SparkIcon';
+import useSnootyMetadata from '../../utils/use-snooty-metadata';
+import { SuspenseHelper } from '../SuspenseHelper';
+
+// lazy load chatbot components
+const Chatbot = lazy(() => import('mongodb-chatbot-ui'));
+const SearchSuggestions = lazy(() => import('./SearchSuggestions'));
 
 const PLACEHOLDER_TEXT = `Search MongoDB Docs or Ask MongoDB AI`;
 
@@ -25,77 +30,24 @@ const inputStyling = css`
   }
 `;
 
-// using content before/after to prevent event bubbling up from lg/search-input/search-result
-// package above gets all text inside node, and sets the value of Input node of all text within search result:
-// https://github.com/mongodb/leafygreen-ui/blob/%40leafygreen-ui/search-input%402.1.4/packages/search-input/src/SearchInput/SearchInput.tsx#L149-L155
-const suggestionStyling = ({ copy }) => css`
-  & > div:before {
-    content: '${copy} "';
-  }
-
-  & > div:after {
-    content: '"';
-  }
-
-  svg:first-of-type {
-    float: left;
-    margin-right: ${theme.size.tiny};
-  }
-
-  padding: ${theme.fontSize.tiny} ${theme.size.medium};
-
-  svg:last-of-type {
-    float: right;
-  }
+const StyledMenu = styled.div`
+  position: absolute;
+  width: 100%;
 `;
 
 const SearchInput = ({ className }) => {
   const [searchValue, setSearchValue] = useState('');
   const [isOpen, setIsOpen] = useState(false);
+  const [chatbotAvailable, setChatbotAvailable] = useState(false);
   const shortcutKeyPressed = useRef(false);
-  const { setInputText, handleSubmit, conversation } = useChatbotContext();
-
-  const createConversation = useCallback(async () => {
-    try {
-      await conversation.createConversation();
-    } catch (e) {
-      console.error('Chatbot not available: ', e);
-    }
-  }, [conversation]);
-
-  const handleClick = useCallback(
-    async (open) => {
-      setInputText(searchValue);
-      if (!conversation.conversationId) {
-        await createConversation();
-      }
-      if (open) {
-        return handleSubmit(searchValue);
-      }
-      window.location.href = `https://mongodb.com/docs/search/?q=${searchValue}`;
-    },
-    [conversation.conversationId, createConversation, handleSubmit, searchValue, setInputText]
-  );
-
-  const SEARCH_SUGGESTIONS = useMemo(
-    () => [
-      {
-        copy: 'Search',
-        onClick: async () => {
-          handleClick(shortcutKeyPressed.current);
-        },
-      },
-      {
-        copy: 'Ask MongoDB AI',
-        onClick: async () => {
-          handleClick(true);
-        },
-        icon: <SparkleIcon glyph={'Sparkle'} />,
-        shortcutIcon: <ShortcutIcon width={30} height={18} />,
-      },
-    ],
-    [handleClick]
-  );
+  const containerRef = useRef();
+  const searchResultsRef = useRef();
+  const metadata = useSnootyMetadata();
+  const { darkMode } = useDarkMode();
+  const CHATBOT_SERVER_BASE_URL =
+    metadata?.snootyEnv === 'dotcomprd'
+      ? 'https://knowledge.mongodb.com/api/v1'
+      : 'https://knowledge.staging.corp.mongodb.com/api/v1';
 
   const keyPressHandler = useCallback((event) => {
     if ((navigator.userAgent.includes('Mac') && event.key === 'Meta') || event.key === 'Control') {
@@ -124,52 +76,31 @@ const SearchInput = ({ className }) => {
   }, [keyPressHandler]);
 
   return (
-    <>
+    <div ref={containerRef} className={cx(inputStyling, className)}>
       <LGSearchInput
         aria-label="Search MongoDB Docs"
         value={searchValue}
         placeholder={PLACEHOLDER_TEXT}
-        className={cx(inputStyling, className)}
         onChange={(e) => {
           setSearchValue(e.target.value);
         }}
-        onFocus={() => {
-          if (!conversation.conversationId) {
-            createConversation();
-          }
-        }}
-      >
-        {isOpen && searchValue.length
-          ? SEARCH_SUGGESTIONS.map((suggestion, i) => {
-              const { copy, icon, onClick } = suggestion;
-
-              return (
-                <SearchResult className={cx(suggestionStyling({ copy, icon }))} key={i} onClick={onClick}>
-                  {suggestion.icon}
-                  {searchValue}
-                  {suggestion.shortcutIcon}
-                </SearchResult>
-              );
-            })
-          : undefined}
-      </LGSearchInput>
-      <ModalView
-        inputBottomText={
-          'This is an experimental generative AI chatbot. All information should be verified prior to use.'
-        }
-        disclaimer={
-          <>
-            <MongoDbLegalDisclosure />
-            <PoweredByAtlasVectorSearch
-              linkStyle="text"
-              className={css`
-                margin-top: 8px;
-              `}
-            />
-          </>
-        }
-      />
-    </>
+      ></LGSearchInput>
+      <StyledMenu ref={searchResultsRef}>
+        <SuspenseHelper>
+          <Chatbot serverBaseUrl={CHATBOT_SERVER_BASE_URL} darkMode={darkMode}>
+            {isOpen && searchValue.length ? (
+              <SearchSuggestions
+                refEl={containerRef}
+                ref={searchResultsRef}
+                searchValue={searchValue}
+                setChatbotAvailable={setChatbotAvailable}
+                chatbotAvailable={chatbotAvailable}
+              />
+            ) : undefined}
+          </Chatbot>
+        </SuspenseHelper>
+      </StyledMenu>
+    </div>
   );
 };
